requirements.txt
amqp==5.1.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-cors-headers==3.7.0
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
loguru==0.7.1
MarkupSafe==2.1.3
numpy==1.24.4
packaging==23.1
pandas==1.5.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pya3==1.0.29
pycryptodome==3.18.0
PyJWT==2.6.0
pymongo==4.3.3
pyotp==2.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3.post1
redis==5.0.0
rel==0.4.9
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.3
